+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [28.0.0] 2020-12-11
+### Changed
+- Updates FHIR Questionnaire definitions to use the observationExtract
+  extension if an observationLinkPeriod is used.
+
 ## [26.0.1] 2020-11-20
 ### Changed
 - The list of update versions is now determined at build time, rather than being